// idcs-multi-user-kmsi/server/types/idcsTypes.d.ts
interface RequestState {
  status?: "success" | "failed" | "pending";
  authnToken?: string;
  requestState: string;
  nextOp?: Array<string>;
  nextAuthFactors?: Array<AuthFactor>;
  //Various Authentication Factor types
  IDP?: IdentityProviderList;
  USERNAME_PASSWORD?: GenericCredentials;
  SMS?: GenericCredentials;
  TOU?: TermsOfUse;
  EMAIL?: EmailCredentials;
  KMSI?: GenericCredentials;
  SECURITY_QUESTIONS?: SecurityQuestionCredentials;
  PUSH?: PushCredentials;
  BYPASSCODE?: GenericCredentials;
  TOTP?: TimebasedOTPCredentials;
  //Used for MFA trust
  trustedDeviceSettings?: TrustedDeviceSettings;
  //KMSI Enabled
  keepMeSignedInEnabled?: boolean;
  //Error message if status is "failed" or "pending"
  cause?: Array<IdcsError>;
  kmsiToken?: string;
  //For when we need to do Reset Passwords or MFA
  EnrolledAccountRecoveryFactorsDetails?: RecoveryFactors;
  //MFA settings define whether we need to enroll
  mfaSettings?:MfaSettings;
  //Display name is used to show MFA factor details
  displayName?:string;
  scenario?:string;
  //UserId is returned for password must change
  userId?:string;
  passwordPolicy?: PasswordPolicy;
}

interface IdcsError {
  message: string;
  code: string;
}
// ...
interface GenericCredentials {
  credentials: Array<string>;
  enrolledDevices?: Array<Device>
}

interface IdentityProviderList extends GenericCredentials {
  configuredIDPs: Array<IdentityProvider>;
}

interface IdentityProvider {
  iconUrl: string;
  idpName: string;
  idpType: "Saml" | "Social"
}

interface TermsOfUse extends GenericCredentials {
  statement: string;
  locale: string;
}

interface EmailCredentials extends GenericCredentials {
  userAllowedToSetRecoveryEmail: string;
  primaryEmailVerified: string;
  primaryEmail: string;
}

interface SecurityQuestionCredentials extends GenericCredentials {
  secQuesSettings: SecurityQuestionSettings;
  securityQuestions: Array<SecurityQuestions>;
}

interface SecurityQuestionSettings {
  minAnswerLength: number;
  maxFieldLength: number;
  numQuestionsToAns: number;
  numQuestionsToSetup: number;
}

interface SecurityQuestions {
  questionId: string;
  text: string;
}

interface PushCredentials {
  qrCode?: QRCode;
  enrolledDevices?: Array<Device>;
}

interface TimebasedOTPCredentials extends GenericCredentials {
  qrCode?: QRCode;
}

interface QRCode {
  imageData: string;
  imageType: string;
  content: string;
}

interface TrustedDeviceSettings {
  trustDurationInDays: number;
}

type AuthFactor = "IDP" | "USERNAME_PASSWORD" | "SMS" | "TOU" | "EMAIL" | "KMSI" | "BYPASSCODE" | "TOTP" | "PUSH" | "SECURITY_QUESTIONS";
// ...
interface RecoveryFactors {
  EMAIL?: RecoveryFactor;
  SMS?: RecoveryFactor;
  enrolledAccRecFactorsList?: Array<"EMAIL"|"SMS">;
}

interface RecoveryFactor {
  credentials: GenericCredentials;
  enrolledDevices: Array<Device>;
}

interface Device {
  deviceId?:string;
  displayName:string;
}

interface MfaSettings {
  enrollmentRequired:boolean;
}

interface MfaCredential {
  [key:string]: string|boolean;
}

interface CountryCodesAllowedValues {
  attrValues: Array<CountryCode>;
}

interface CountryCode {
  label: string;
  value: string;
}

interface PasswordPolicy {
  name?:string;
  id?:string;
  minLength?: number;
  maxLength?: number;
  minLowerCase?: number;
  minUpperCase?: number;
  minAlphas?: number
  minNumerals?: number;
  minAlphaNumerals?: number;
  minSpecialChars?: number;
  minUniqueChars?: number;
  maxRepeatedChars?: number;
  startsWithAlphabet?: boolean;
  userNameDisallowed?: boolean;
  firstNameDisallowed?: boolean;
  lastNameDisallowed?: boolean; 
  disallowedChars?: string;
  requiredChars?:string;
  numPasswordsInHistory?: number;
}